# shared/api-contracts/openapi.yaml
...
openapi: "3.1.0"
info:
  title: "TTBS Customer API"
  description: "Customer-facing API contract for the train ticket booking system."
  version: "1.0"
servers:
- url: "/api"
  description: "Versioned customer API root"
security:
- bearerAuth: []
...
paths:
...
  /api/{version}/trains:
    get:
      tags:
      - "Trains"
      summary: "List trains"
      operationId: "getTrains"
      parameters:
      - name: "request"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/GetTrainsRequest"
      responses:
        "200":
          description: "Paged train catalog"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseTrainResponse"
        "400":
          description: "Invalid pagination parameters"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
  /api/{version}/trains/{trainId}/seats:
    get:
      tags:
      - "Trains"
      summary: "List seats for a train"
      operationId: "getTrainSeats"
      parameters:
      - name: "trainId"
        in: "path"
        description: "Train identifier"
        required: true
        schema:
          type: "string"
          format: "uuid"
      - name: "request"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/GetSeatsRequest"
      responses:
        "200":
          description: "Paged seats for the train"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseSeatResponse"
        "400":
          description: "Invalid pagination parameters"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
...
  /api/{version}/trains/{id}:
    get:
      tags:
      - "Trains"
      summary: "Get a train by id"
      operationId: "getTrain"
      parameters:
      - name: "id"
        in: "path"
        description: "Train identifier"
        required: true
        schema:
          type: "string"
          format: "uuid"
      - name: "request"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/GetTrainByIdRequest"
      responses:
        "200":
          description: "Train detail"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/TrainResponse"
        "404":
          description: "Train not found"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
...
  /api/{version}/scheduled-trips/{scheduledTripId}/seats/available:
    get:
      tags:
      - "Trains"
      summary: "List available seats for a scheduled trip"
      operationId: "getAvailableSeats"
      parameters:
      - name: "scheduledTripId"
        in: "path"
        description: "Scheduled trip identifier"
        required: true
        schema:
          type: "string"
          format: "uuid"
      - name: "request"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/GetAvailableSeatsRequest"
      responses:
        "200":
          description: "Paged available seats"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseSeatResponse"
        "400":
          description: "Invalid seat filter or pagination parameters"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "404":
          description: "Scheduled trip not found"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
  /api/{version}/scheduled-trips/{scheduledTripId}/coach-seats:
    get:
      tags:
      - "Trains"
      summary: "Get the seat map for a scheduled trip"
      operationId: "getCoachSeatMap"
      parameters:
      - name: "scheduledTripId"
        in: "path"
        description: "Scheduled trip identifier"
        required: true
        schema:
          type: "string"
          format: "uuid"
      - name: "request"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/GetCoachSeatMapRequest"
      responses:
        "200":
          description: "Coach seat map for the scheduled trip"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/CoachSeatMapResponse"
        "400":
          description: "Invalid coach selection parameters"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "404":
          description: "Scheduled trip not found"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
...
components:
  schemas:
    FailData:
      description: "Shared payload used inside JSend fail responses."
      properties:
        message:
          type: "string"
          description: "Human-readable summary of the failure."
        code:
          $ref: "#/components/schemas/ErrorCode"
        errors:
          type: "array"
          items:
            $ref: "#/components/schemas/Violation"
    ErrorCode:
      type: "string"
      description: "Machine-readable error code carried by shared fail responses."
      enum:
      - "VALIDATION_ERROR"
      - "CURSOR_INVALID"
      - "ACCESS_DENIED"
      - "USER_EMAIL_ALREADY_EXISTS"
      - "USER_INVALID_CREDENTIALS"
      - "USER_INVALID_REFRESH_TOKEN"
      - "USER_NOT_FOUND"
      - "USER_CANNOT_BULK_DELETE_SELF"
      - "USER_HAS_ACTIVE_BOOKINGS"
      - "BOOKING_NOT_FOUND"
      - "BOOKING_CANNOT_CONFIRM"
      - "BOOKING_ALREADY_CANCELLED"
      - "TRAIN_NUMBER_ALREADY_EXISTS"
      - "TRAIN_NOT_FOUND"
      - "TRAIN_IN_USE"
      - "SEAT_NOT_FOUND"
      - "SEAT_NUMBER_ALREADY_EXISTS"
      - "SEAT_NOT_AVAILABLE"
      - "SEAT_IN_USE"
      - "SCHEDULED_TRIP_NOT_FOUND"
      - "COACH_NOT_FOUND"
      - "COACH_CAR_NUMBER_ALREADY_EXISTS"
      - "COACH_TRAIN_NOT_FOUND"
      - "COACH_IN_USE"
      - "COACH_CAR_NUMBERS_ALREADY_EXIST"
      - "COACH_DUPLICATE_CAR_NUMBERS_IN_REQUEST"
      - "SEAT_NUMBERS_ALREADY_EXIST"
      - "SEAT_DUPLICATE_SEAT_NUMBERS_IN_REQUEST"
      - "ROUTE_TEMPLATE_NOT_FOUND"
      - "STATION_NOT_FOUND"
      - "STATION_CODE_ALREADY_EXISTS"
      - "STATION_IN_USE"
      - "PAYMENT_NOT_FOUND"
      - "PAYMENT_ALREADY_PROCESSED"
      - "PAYMENT_REFUND_FAILED"
    Violation:
      description: "Field-level validation issue inside a fail response."
      properties:
        field:
          type: "string"
          description: "Request field that failed validation."
        message:
          type: "string"
          description: "Human-readable validation message."
        code:
          $ref: "#/components/schemas/ViolationCode"
    ViolationCode:
      type: "string"
      description: "Machine-readable validation category attached to a field violation."
      enum:
      - "REQUIRED"
      - "INVALID_FORMAT"
      - "TOO_SHORT"
      - "TOO_LONG"
      - "INVALID_VALUE"
    JsendFailResponse:
      type: "object"
      description: "Shared JSend fail envelope used for validation and domain errors."
      properties:
        status:
          type: "string"
          enum:
          - "fail"
        data:
          $ref: "#/components/schemas/FailData"
    JsendErrorResponse:
      type: "object"
      description: "Shared JSend error envelope used for unexpected technical failures."
      properties:
        status:
          type: "string"
          enum:
          - "error"
        message:
          type: "string"
          description: "Human-readable technical failure message."
...
    JsendResponseObject:
      type: "object"
      description: "Runtime JSend response envelope used by the API implementation."
      properties:
        status:
          type: "string"
          description: "JSend status value."
          enum:
          - "success"
          - "fail"
          - "error"
        data:
          description: "Payload returned for the response status."
        message:
          type: "string"
          description: "Human-readable technical error message for error responses."
...
    GetTrainsRequest:
      type: "object"
      description: "Offset-based pagination query for trains."
      properties:
        page:
          type: "integer"
          format: "int32"
          description: "Zero-based page index."
          example: 0
          minimum: 0
        size:
          type: "integer"
          format: "int32"
          description: "Number of trains per page."
          example: 20
          maximum: 100
          minimum: 1
    GetSeatsRequest:
      type: "object"
      description: "Offset-based pagination query for seats belonging to a train."
      properties:
        page:
          type: "integer"
          format: "int32"
          description: "Zero-based page index."
          example: 0
          minimum: 0
        size:
          type: "integer"
          format: "int32"
          description: "Number of seats per page."
          example: 20
          maximum: 100
          minimum: 1
...
    GetTrainByIdRequest:
      description: "No additional query parameters are accepted when fetching a train."
...
    GetAvailableSeatsRequest:
      type: "object"
      description: "Offset-based pagination query for currently available seats on\
        \ a scheduled trip."
      properties:
        page:
          type: "integer"
          format: "int32"
          description: "Zero-based page index."
          example: 0
          minimum: 0
        size:
          type: "integer"
          format: "int32"
          description: "Number of seats per page."
          example: 20
          maximum: 100
          minimum: 1
    GetCoachSeatMapRequest:
      type: "object"
      description: "Offset-based pagination query for a scheduled trip coach seat\
        \ map."
      properties:
        page:
          type: "integer"
          format: "int32"
          description: "Zero-based page index."
          example: 0
          minimum: 0
        size:
          type: "integer"
          format: "int32"
          description: "Number of seats per page."
          example: 20
          maximum: 100
          minimum: 1
...
    PageResponseTrainResponse:
      description: "Offset-based pagination wrapper."
      properties:
        content:
          type: "array"
          items:
            $ref: "#/components/schemas/TrainResponse"
        page:
          type: "integer"
          format: "int32"
          description: "Zero-based page index."
          example: 0
          minimum: 0
        size:
          type: "integer"
          format: "int32"
          description: "Requested page size."
          example: 20
          minimum: 1
        hasNext:
          type: "boolean"
          description: "Whether a subsequent page exists."
        hasPrevious:
          type: "boolean"
          description: "Whether a previous page exists."
        total:
          type: "integer"
          format: "int64"
          description: "Total number of matching items across all pages."
          example: 125
          minimum: 0
    TrainResponse:
      description: "Train summary."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Train identifier."
        trainNumber:
          type: "string"
          description: "Train number visible to operators and customers."
          example: "SE1"
        name:
          type: "string"
          description: "Train display name."
          example: "Reunification Express"
        totalSeats:
          type: "integer"
          format: "int32"
          description: "Total number of seats across the train."
          example: 320
        createdAt:
          type: "string"
          format: "date-time"
          description: "Train creation timestamp."
    PageResponseSeatResponse:
      description: "Offset-based pagination wrapper."
      properties:
        content:
          type: "array"
          items:
            $ref: "#/components/schemas/SeatResponse"
        page:
          type: "integer"
          format: "int32"
          description: "Zero-based page index."
          example: 0
          minimum: 0
        size:
          type: "integer"
          format: "int32"
          description: "Requested page size."
          example: 20
          minimum: 1
        hasNext:
          type: "boolean"
          description: "Whether a subsequent page exists."
        hasPrevious:
          type: "boolean"
          description: "Whether a previous page exists."
        total:
          type: "integer"
          format: "int64"
          description: "Total number of matching items across all pages."
          example: 125
          minimum: 0
    SeatResponse:
      description: "Seat summary."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Seat identifier."
        coachId:
          type: "string"
          format: "uuid"
          description: "Parent coach identifier."
        seatNumber:
          type: "string"
          description: "Seat label shown to customers."
          example: "12A"
        createdAt:
          type: "string"
          format: "date-time"
          description: "Seat creation timestamp."
...
    CoachSeatMapResponse:
      description: "Seat map for a coach on a scheduled trip."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Coach identifier."
        carNumber:
          type: "integer"
          format: "int32"
          description: "Coach car number within the train."
          example: 5
        totalSeats:
          type: "integer"
          format: "int32"
          description: "Total seats in the coach."
          example: 64
        seats:
          type: "array"
          items:
            $ref: "#/components/schemas/Seat"
    Seat:
      description: "Seat state inside a coach seat map."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Seat identifier."
        seatNumber:
          type: "string"
          description: "Seat label shown to customers."
          example: "12A"
        status:
          type: "string"
          description: "Current seat availability status."
          enum:
          - "AVAILABLE"
          - "HELD"
          - "BOOKED"
          - "CANCELLED"
...
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"